Refactor project directory
