--- src/components/About.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const About = () => {
-  return(
-    <div>lorem ipsum</div>
-  );
-};
-
-export default About;
