--- src/components/avocadoro/task/Task.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import TaskCard from './TaskCard';
-
-const Task = () => {
-  return(
-    <div>
-      <span>Task</span>
-      <TaskCard/>
-      <TaskCard/>
-      <TaskCard/>
-    </div>
-  );
-};
-
-export default Task;
