+++ src/components/dashboard/timer/TimerInfo.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const TimerInfo = (props) => {
+  return (
+    <div>
+      TimerInfo
+    </div>
+  );
+};
+
+export default TimerInfo;
